+++ src/main/java/yggc/service/iss/IndexNewsService.java
@@ -24,5 +24,5 @@ public interface IndexNewsService {
 	* @param @return      
 	* @return List<Article>
 	 */
-	List<Article> selectNewsByArticleTypeId(Long id);
+	List<Article> selectNewsByArticleTypeId(Integer id);
 }
